% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbo_html_report.R
\name{create_summary_report}
\alias{create_summary_report}
\title{Create a report summarising results of screen}
\usage{
create_summary_report(
  optimisations,
  outdir = getwd(),
  prefix = "screen_results"
)
}
\arguments{
\item{optimisations}{an \code{\link{OptimisationResultCollection}} object produced by \code{\link[=screen_molecules]{screen_molecules()}}}

\item{outdir}{Directory where the HTML report should be written.
Defaults to the current working directory. The final report will be saved
as \verb{\{prefix\}.html} inside this folder.}

\item{prefix}{A character string used as the base filename for the report.
The resulting output file will be named \verb{\{prefix\}.html}. Defaults to
\code{"screen_results"}.}
}
\value{
invisibly returns NULL (this function is run for its side effects)
}
\description{
Create a report summarising results of screen
}
