% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maths.R
\name{rotate_vector_into_a_plane}
\alias{rotate_vector_into_a_plane}
\title{Compute the rotation needed to align a 3D vector with a plane}
\usage{
rotate_vector_into_a_plane(v, plane_normal)
}
\arguments{
\item{v}{numeric(3) Vector to rotate.}

\item{plane_normal}{numeric(3) Plane normal (not necessarily unit length).}
}
\value{
list with elements \code{axis} (unit vector) and \code{angle} (radians).
}
\description{
Calculates the minimal rotation that moves vector \code{v} so it lies in the
plane with normal \code{plane_normal}. Returns a list containing the rotation
axis (unit vector) and rotation angle (radians).
}
\examples{
rotate_vector_into_a_plane(c(1,2,3), c(0,1,-1))
}
