% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbo-package.R
\docType{package}
\name{symbo-package}
\alias{symbo}
\alias{symbo-package}
\title{symbo: Symmetry-Based Screening of Pairwise Molecular Assemblies}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Symbo takes two molecular structures and assesses (based on their symmetry properties) the geometric feasability of forming different higher-order structures like face/edge capped cubes or tetrahedra.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/selkamand/symbo}
  \item \url{https://selkamand.github.io/symbo/}
  \item Report bugs at \url{https://github.com/selkamand/symbo/issues}
}

}
\author{
\strong{Maintainer}: Sam El-Kamand \email{selkamand@ccia.org.au} (\href{https://orcid.org/0000-0003-2270-8088}{ORCID})

Authors:
\itemize{
  \item Matthew Wallis (\href{https://orcid.org/0009-0004-8113-9342}{ORCID})
}

}
\keyword{internal}
