% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbo_html_report.R
\name{create_summary_report}
\alias{create_summary_report}
\title{Create a report summarising results of screen}
\usage{
create_summary_report(
  optimisations,
  outdir = getwd(),
  prefix = "screen_results"
)
}
\arguments{
\item{optimisations}{an \code{\link{OptimisationResultCollection}} object produced by \code{\link[=screen_molecules]{screen_molecules()}}}
}
\value{
invisibly returns NULL (this function is run for its side effects)
}
\description{
Create a report summarising results of screen
}
