% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbo-package.R
\docType{package}
\name{symbo-package}
\alias{symbo}
\alias{symbo-package}
\title{symbo: What the Package Does (One Line, Title Case)}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/selkamand/symbo}
  \item \url{https://selkamand.github.io/symbo/}
  \item Report bugs at \url{https://github.com/selkamand/symbo/issues}
}

}
\author{
\strong{Maintainer}: Sam El-Kamand \email{selkamand@gmail.com} (\href{https://orcid.org/0000-0003-2270-8088}{ORCID}) [copyright holder]

}
\keyword{internal}
